feat:HeroSection design
